Upload Files
Live Project Summary.

Interning and working on the full feature application, TravelScrape, was challenging and rewarding.  I worked in a team of 8 developers all working on different stories.  We utilized Agile methodology of Project Lifecycle Management which incorporated Scrum framework with daily sprints.  I participated in daily stand ups, as well as, weekly retrospectives.

Working in a legacy codebase, I was able to jump in and familiarize myself with the project.  I worked on a front end project that incorporated the use of HTML, CSS, and Javascript.  All work was completed in Django.  My work had to be adjusted due to changes in other stories, in which I benefitted by debugging and continuing to incorporate my work in the project.  I also further extended my story by adding a useful feature for the end user.  On the backend, I initiated the creation of a new app within the application.  This was challenging and exciting as I traversed my way through Django and Git.  

Attached are Below are snippets of the frontend and backend work I completed during my internship.  

Home page interactive buttons (with the added feature of hide when menubar is active).

Create a US Cities app.